Add test with noauth for s3tokens and ec2tokens
The s3tokens and ec2tokens endpoint is a public API
that does not require authentication to be called but
that is what the testing does today.

This duplicates the "good request" tests by also
testing with noauth set so that we don't send a
token in the request, this can prevent bugs where
we would unintentionally require auth for these
endpoints.

Change-Id: Ibbde313ac8bcf3187c139bbd0840702f229534d0
(cherry picked from commit bd70653a2445fe1b9628ff1f92cbc58f2c9f4d70)